Make contract fixtures linkable outside the test tree (issue #51)
Extract equilibrium_seed.h and the GTest-free subset of al_contract.h's
generic typed round-trip helpers it needs (DType, write_data, read_data,
element_count) into a new top-level fixtures/ directory, wired as an
INTERFACE library (al_contract_fixtures) added unconditionally in the
root CMakeLists.txt, independent of AL_BUILD_TESTS. This lets a future
benchmarks/ target (PRD #50) link the exact same workload the contract
suite pins, with no fixture drift and no duplication.
tests/contract/al_contract.h now includes the shared header instead of
redefining those symbols; contract_tests links the new target so
unqualified #include "equilibrium_seed.h" in the existing test files
keeps resolving.

An opt-in throwaway executable (AL_BUILD_FIXTURE_SMOKE, off by default)
proves the fixture builds and round-trips against the Memory backend
from a target outside tests/, with no GoogleTest link.
