+++ tests/contract/al_contract.h
@@ -12,6 +12,8 @@
 #ifndef AL_CONTRACT_H
 #define AL_CONTRACT_H
 
+#include "al_contract_abi.h"
+
 #include <al_lowlevel.h>
 #include <al_const.h>
 
@@ -20,8 +22,6 @@
 #include <atomic>
 #include <cerrno>
 #include <chrono>
-#include <complex>
-#include <cstddef>
 #include <cstdlib>
 #include <cstring>
 #include <filesystem>
@@ -345,30 +345,12 @@ inline al_status_t read_char_array(int ctx, const char* field,
 // itself the proof of the version-free property. This block generalizes the
 // two hand-written scalar/char helpers above to the whole
 // datatype × shape space.
-
-// --- datatype descriptor: element C++ type -> ABI datatype id + label -------
-template <class T>
-struct DType;
-template <>
-struct DType<char> {
-    static constexpr int         value = CHAR_DATA;
-    static constexpr const char* name  = "CHAR";
-};
-template <>
-struct DType<int> {
-    static constexpr int         value = INTEGER_DATA;
-    static constexpr const char* name  = "INTEGER";
-};
-template <>
-struct DType<double> {
-    static constexpr int         value = DOUBLE_DATA;
-    static constexpr const char* name  = "DOUBLE";
-};
-template <>
-struct DType<std::complex<double>> {
-    static constexpr int         value = COMPLEX_DATA;
-    static constexpr const char* name  = "COMPLEX";
-};
+//
+// DType/write_data/read_data/element_count (the subset equilibrium_seed.h also
+// needs) live in al_contract_abi.h (issue #51), included above, so a build
+// target outside tests/contract/ can link them without GoogleTest. The
+// synthetic-buffer generator and its shape helper, used exclusively by this
+// suite's own matrix tests, stay here.
 
 // The core's "this value means absent" sentinels (src/al_lowlevel.cpp:41-44).
 // A *scalar* write whose value equals its sentinel is silently dropped by
@@ -420,67 +402,12 @@ inline std::vector<int> shape_for_rank(int rank) {
     return s;
 }
 
-inline std::size_t element_count(const std::vector<int>& shape) {
-    std::size_t n = 1;
-    for (int d : shape) n *= static_cast<std::size_t>(d);
-    return n;
-}
-
 template <class T>
 std::vector<T> synth_buffer(std::size_t n) {
     std::vector<T> v;
     v.reserve(n);
     for (std::size_t i = 0; i < n; ++i) v.push_back(synth_value<T>(i));
     return v;
-}
-
-// --- generic write ----------------------------------------------------------
-// al_write_data takes a non-const void* and int* size but does not mutate
-// them for a write; the const_casts keep the caller's data/shape const.
-template <class T>
-al_status_t write_data(int ctx, const char* field, const std::vector<int>& shape,
-                       const std::vector<T>& data) {
-    const int dim  = static_cast<int>(shape.size());
-    int*      size = dim ? const_cast<int*>(shape.data()) : nullptr;
-    return al_write_data(ctx, field, "", const_cast<T*>(data.data()),
-                         DType<T>::value, dim, size);
-}
-
-// --- generic read -----------------------------------------------------------
-// Honors the two ownership regimes the ABI uses (mirrors testlowlevel.cpp):
-//   dim == 0 : the core writes into the caller's buffer; nothing to free.
-//   dim >= 1 : the core malloc's the buffer and returns it via *data; the
-//              caller owns it, so we copy out and free.
-// On success out_shape is the shape the core reported (scalar -> empty) and
-// out_data holds element_count(out_shape) elements.
-template <class T>
-al_status_t read_data(int ctx, const char* field, int expected_rank,
-                      std::vector<int>* out_shape, std::vector<T>* out_data) {
-    int size[MAXDIM] = {0};
-
-    if (expected_rank == 0) {
-        T     scalar{};
-        void* buf = &scalar;
-        al_status_t s =
-            al_read_data(ctx, field, "", &buf, DType<T>::value, 0, size);
-        if (s.code == 0) {
-            out_shape->clear();
-            out_data->assign(1, scalar);
-        }
-        return s;
-    }
-
-    void*       buf = nullptr;
-    al_status_t s   = al_read_data(ctx, field, "", &buf, DType<T>::value,
-                                   expected_rank, size);
-    if (s.code == 0 && buf != nullptr) {
-        out_shape->assign(size, size + expected_rank);
-        const std::size_t n = element_count(*out_shape);
-        T*                p = static_cast<T*>(buf);
-        out_data->assign(p, p + n);
-    }
-    free(buf);
-    return s;
 }
 
 }  // namespace al_contract
